add-moves-tasks-and-fix-bugs

+++ src/redux-store/tasks/tasksOperations.js
@@ -1,17 +1,23 @@
 import axios from 'axios';
 import { createAsyncThunk } from '@reduxjs/toolkit';
-// import { token } from 'redux-store/AuthOperations/AuthOperations';
+import swal from 'sweetalert';
+import { token } from 'redux-store/AuthOperations/AuthOperations';
 
-// axios.defaults.baseURL = 'https://goose-track-ity9.onrender.com/api/';
+axios.defaults.baseURL = 'https://goose-track-ity9.onrender.com/api/';
 
 export const fetchTasksMonth = createAsyncThunk(
   'tasks/fetchAllMonth',
   async ({ date }, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const persistedToken = state.auth.token;
+    if (persistedToken === null) {
+      return thunkAPI.rejectWithValue('No valid token');
+    }
     try {
+      token.set(persistedToken);
       const response = await axios.get('/tasks/month', {
         params: { date },
       });
-
       return response.data;
     } catch (error) {
       return thunkAPI.rejectWithValue(error.message);
@@ -22,7 +28,14 @@ export const fetchTasksMonth = createAsyncThunk(
 export const fetchTasksDay = createAsyncThunk(
   'tasks/fetchAllDay',
   async ({ date }, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const persistedToken = state.auth.token;
+
+    if (persistedToken === null) {
+      return thunkAPI.rejectWithValue('No valid token');
+    }
     try {
+      token.set(persistedToken);
       const response = await axios.get('/tasks/day', {
         params: { date },
       });
@@ -46,11 +59,12 @@ export const addTask = createAsyncThunk(
 
     try {
       const response = await axios.post('/tasks', body);
-      console.log(response);
       return response.data;
-    } catch (error) {
-      console.log(error);
-      return thunkAPI.rejectWithValue(error.message);
+    } catch (e) {
+      if (e.response.status === 400 || e.response.status === 409) {
+        throw new Error(swal('Error!', e.response.data.message, 'error'));
+      }
+      return thunkAPI.rejectWithValue(e.message);
     }
   }
 );
@@ -58,8 +72,13 @@ export const addTask = createAsyncThunk(
 export const editTask = createAsyncThunk(
   'tasks/editTask',
   async (task, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const persistedToken = state.auth.token;
+
+    if (persistedToken === null) {
+      return thunkAPI.rejectWithValue('No valid token');
+    }
     const { _id, ...edit } = task;
-    console.log('task***', task);
     try {
       const response = await axios.patch(`/tasks/${_id}`, edit);
       return response.data;
@@ -72,6 +91,12 @@ export const editTask = createAsyncThunk(
 export const deleteTask = createAsyncThunk(
   'tasks/deleteTask',
   async (taskId, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const persistedToken = state.auth.token;
+
+    if (persistedToken === null) {
+      return thunkAPI.rejectWithValue('No valid token');
+    }
     try {
       await axios.delete(`/tasks/${taskId}`);
       return taskId;
